changed to allow for sending to just organizer
