+++ code/noRAM/app/src/main/AndroidManifest.xml
@@ -79,7 +79,7 @@
                 android:resource="@xml/file_paths" />
         </provider>
         <service
-            android:name=".PushNotificationService"
+            android:name=".model.PushNotificationService"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
