+++ code/noRAM/app/src/main/java/com/example/noram/OrganizerEventNotificationsActivity.java
@@ -85,7 +85,7 @@ public class OrganizerEventNotificationsActivity extends AppCompatActivity {
             // Send notification
             String title = editTitle.getText().toString();
             String content = editContent.getText().toString();
-            MainActivity.pushService.sendNotification(title, content, event);
+            MainActivity.pushService.sendNotification(title, content, event, false);
             finish();
         });
