+++ code/noRAM/app/src/main/java/com/example/noram/model/PushNotificationService.java
@@ -4,7 +4,7 @@ Outstanding Issues:
 - Currently only sends notifications in app to the main activity
  */
 
-package com.example.noram;
+package com.example.noram.model;
 
 import static android.content.ContentValues.TAG;
 
@@ -13,6 +13,7 @@ import android.util.Log;
 
 import androidx.annotation.NonNull;
 
+import com.example.noram.MainActivity;
 import com.example.noram.model.Event;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
@@ -21,6 +22,7 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.IOException;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
@@ -77,14 +79,16 @@ public class PushNotificationService extends FirebaseMessagingService {
      * @param title the title of the notification
      * @param data the data of the notification
      * @param event the event to send the notification to
+     * @param isOrganizer a boolean to check if we wish to only send the notification to the organizer
      */
-    public void sendNotification(String title, String data, Event event) {
+    public void sendNotification(String title, String data, Event event, Boolean isOrganizer) {
 
-        // Get the list of attendees
-        Set<String> attendeeList = new HashSet<>(event.getCheckedInAttendees()); // cast to Set<String> so no duplicate attendees
+        Set<String> attendeeList;
 
-        if (attendeeList == null) {
-            throw new IllegalArgumentException("Attendee list is null");
+        if (isOrganizer) {
+            attendeeList = Collections.singleton((event.getOrganizerId()));
+        } else {
+            attendeeList = new HashSet<>(event.getCheckedInAttendees()); // cast to Set<String> so no duplicate attendees
         }
 
         // Send a notification to each attendee
